feat: webpack improvements (#4)
This will allow omitting the `entry` in the widget definition file. This is because we'll default to `index.js`.

BREAKING CHANGE: It also leverages webpack to create the `window.renderMyWidget = ...`. One only needs to export a function from now on. This is a breaking change, since render files need an export now.